// web-app/lib/db/prisma/generated/models/Campaign.ts
export type CampaignModel = runtime.Types.Result.DefaultSelection<Prisma.$CampaignPayload>
// ...
export type $CampaignPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Campaign"
  objects: {
    brand: Prisma.$UserPayload<ExtArgs>
    payoutRates: Prisma.$PayoutRatePayload<ExtArgs>[]
    participations: Prisma.$ParticipationPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    brandId: string
    title: string
    escrowAddress: string
    budgetTotal: runtime.Decimal
    yellowChannelId: string | null
    status: $Enums.CampaignStatus
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["campaign"]>
  composites: {}
}